+++ src/main/java/net/more_rpg_classes/client/particle/MoreParticles.java
@@ -27,6 +27,7 @@ public class MoreParticles {
     public static final SimpleParticleType STONE_PARTICLE = FabricParticleTypes.simple();
     public static final SimpleParticleType WIND_VACUUM = FabricParticleTypes.simple();
     public static final SimpleParticleType SMALL_GUST = FabricParticleTypes.simple();
+    public static final SimpleParticleType GAS_CLOUD = FabricParticleTypes.simple();
 
 
     public static void register(){
@@ -49,6 +50,7 @@ public class MoreParticles {
         Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "stone_particle"), STONE_PARTICLE);
         Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "wind_vacuum"), WIND_VACUUM);
         Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "small_gust"), SMALL_GUST);
+        Registry.register(Registries.PARTICLE_TYPE, Identifier.of(MRPGCMod.MOD_ID, "gas_cloud"), GAS_CLOUD);
     }
 
 }