+++ src/main/java/net/more_rpg_classes/effect/FrozenSolidEffect.java
@@ -1,7 +1,6 @@
 package net.more_rpg_classes.effect;
 
 import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 
@@ -28,12 +27,7 @@ public class FrozenSolidEffect extends StatusEffect {
         return true;
     }
 
-
-    public void onApplied(LivingEntity entity, AttributeContainer attributes, int amplifier) {
-        entity.setFrozenTicks(entity.getFrozenTicks() + 40);
-    }
-
-    public void onRemoved(LivingEntity entity, AttributeContainer attributes, int amplifier) {
-        entity.setFrozenTicks(entity.getFrozenTicks() + 200);
+    public static void onRemove(LivingEntity entity) {
+        entity.setFrozenTicks(entity.getFrozenTicks() + 100);
     }
 }